+++ src/OTvk.java
@@ -1,8 +1,9 @@
+import java.math.BigInteger;
 
 public class OTvk {
-	public final int x, a;
+	public final BigInteger x, a;
 	
-	public OTvk(int x, int a){
+	public OTvk(BigInteger x, BigInteger a){
 		this.x = x;
 		this.a = a;
 	}
