+++ src/Pair.java
@@ -1,9 +1,10 @@
+import java.math.BigInteger;
 
 public class Pair {
 	
-	public final int x1,x2;
+	public final BigInteger x1,x2;
 	
-	public Pair(int x1, int x2){
+	public Pair(BigInteger x1, BigInteger x2){
 		this.x1 = x1;
 		this.x2 = x2;
 	}
